Nits: move tests to their own module, pub(crate) instead of pub
